pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "readwise-tools"
version = "0.1.0"
description = "Tools for integrating various services with Readwise"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "requests",
    "readwise",
    "python-dotenv",
    "pycketcasts",
]

[project.scripts]
mastodon-to-readwise = "readwise_tools.mastodon_to_readwise:main"
pocketcasts-to-readwise = "readwise_tools.pocketcasts_to_readwise:main"
readwise-open-links = "readwise_tools.readwise_open_links:main"
readwise-tag-later = "readwise_tools.readwise_tag_later:main"

[project.urls]
Homepage = "https://github.com/christofdamian/readwise-tools"
Repository = "https://github.com/christofdamian/readwise-tools"